Fix race condition by labeling port messages
The race can be triggered by enabling interrupts and toggling the gpio
quickly externally. Then try calling read/1. Eventually, the read/1
receive will match on a notification instead of the result. The result
will then be interpreted as a notification and fail to match. It will be
either a 0 or 1 instead of a tuple.