Allows concurrent calls to SPI
When sending two calls to one SPI connection the connection would
confuse the second call with the response to the first. This is came
about because the `receive` block had a catch all for errors, but that
was catching the new incoming request. Removing the catch all allows
both requests to process.

Amos King @adkron <amos@binarynoggin.com>