Allow :start_value in GPIO.start_link opts to explicitly set value
This helps in the cases that one might want to have the initial starting value of a pin be `1`, especially when using this under supervision. Or just if one wants to ensure a specific value is set on port start.

A minor use-case, but definitely helpful.

I debated using something like `start_link(16, {:output, 1})` vs putting this in the opts, but decided to go this route for more explicity options and see if there are opinions otherwise.